+++ src/com/landawn/abacus/util/ImmutableSortedMap.java
@@ -129,6 +129,8 @@ public class ImmutableSortedMap<K, V> extends ImmutableMap<K, V> implements Sort
     public static <K, V> ImmutableSortedMap<K, V> of(final SortedMap<? extends K, ? extends V> sortedMap) {
         if (sortedMap == null) {
             return empty();
+        } else if (sortedMap instanceof ImmutableSortedMap) {
+            return (ImmutableSortedMap<K, V>) sortedMap;
         }
 
         return new ImmutableSortedMap<>(sortedMap);
@@ -137,11 +139,9 @@ public class ImmutableSortedMap<K, V> extends ImmutableMap<K, V> implements Sort
     public static <K, V> ImmutableSortedMap<K, V> copyOf(final SortedMap<? extends K, ? extends V> sortedMap) {
         if (N.isNullOrEmpty(sortedMap)) {
             return empty();
-        } else if (sortedMap instanceof ImmutableSortedMap) {
-            return (ImmutableSortedMap<K, V>) sortedMap;
         }
 
-        return of(new TreeMap<>(sortedMap));
+        return new ImmutableSortedMap<>(new TreeMap<>(sortedMap));
     }
 
     @Override