+++ CHANGES.md
@@ -11,6 +11,8 @@
 
 * Add `Optional<T> SQLExecutor.geT(...)`.
 
+* Rename `ImmutableSet.of(E... a)` to `ImmutableSet.from(E... a)`.
+
 * Improvements and bug fix.